colorspace: Add WIP protocol for colorspace stuff
TODO: Figure out which chromacities / transfer functions are
needed/useful. Alternative would be to pass the primaries and white
point as xyY coordinates, but then the compositor might have
to reverse map that back into an enum for some hardware implementations
and whanot. Passing the transfer function is a parametrized curve
doesn't seem viable on account of ST 20854 PQ and HLG. Normal gamma
(with or without the linear segment) could be parametrized.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>