drm-backend: Prepare connector's color state to handle HDR view
This patch adds initial code to handle a HDR view in DRM backend,
while plane assignment. In order to handle HDR playback, the first
step is to identify a HDR view, by identifying a HDR surface. ie,
if any of the surface comes with HDR metadata, its a HDR surface,
which makes this view a HDR view.

Now, once we identify an HDR view, we have to do following:
- Tone mapping:
	- If output monitor is HDR monitor, we have to prepare output
 	HDR metadata, set that in connector's HDR metadata property,
  	and tone map all the surfaces as per output metadata:
  	SDR surfaces -> tone map SDR to HDR (S2H) -> HDR output
  	HDR surfaces -> tone map HDR to HDR (H2H) -> HDR output

	- If output monitor is a SDR monitor, we have to reset the HDR
  	output color property (if set), and tone map the HDR surfaces
  	SDR surface -> no change -> SDR output
	HDR surface -> H2S tone mapping -> SDR output

- Set output colorspace
	If output is HDR -> set output colorspace as REC2020/DCIP3
	If output is SDR -> set output colorspace as REC709/default

This patch prepares the required calculations, and prepares the
connector's color state to be commited, and reflect these changes
in current atomic commit.

Signed-off-by: Shashank Sharma <shashank.sharma@intel.com>