drm-backend: Handle HDR view
This patch tries to handle HDR view, by creating an output state
using display engine's overlay planes. As we are adding limited
HDR playback support, this function restricts the buffers to be
displayed, like:
- This functions expects only one HDR surface to be present in the
  view, and it cosiders the first HDR surface for all calculations.
- Weston supports only DMA buffers to be sent to overlays, so this
  function considers only DMA buffers for state creation.

If the state creation is successful, it commits the output state
to weston, by moving the view to planes. If not, like all other
overlay failures, it falls to GL composition, suppoprt for which
will be added later in this series.

Signed-off-by: Shashank Sharma <shashank.sharma@intel.com>