drm_backend: Add plane color correction properties
In order to show accurate display, and to be able to show
SDR and HDR buffers at the same time, we need to add some
color corrections to the buffers assigned to the overlay
planes. This patch adds following plane properies for the
same:
- plane level degamma lut size
- plane level degamma
- plane level CSC
- plane level gamma lut size
- plane level gamma

This patch also adds a plane color state infrastructure,
to reflect the current color correction status for each
of these planes.

Signed-off-by: Shashank Sharma <shashank.sharma@intel.com>